requirements.txt
absl-py==1.0.0
altgraph==0.17.2
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
bidict==0.21.4
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.2
Flask==2.0.3
flask-ngrok==0.0.25
Flask-SocketIO==5.1.1
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.1
imutils==0.5.4
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
parso==0.8.3
pathspec==0.9.0
pefile==2021.9.3
pickleshare==0.7.5
pigar==0.10.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.0
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
Pygments==2.11.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pyngrok==5.1.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
yarg==0.1.9
zipp==3.7.0
